refactor: reorganize tests and add playback sync features
- Reorganize test structure into src/__tests__/unit and src/__tests__/e2e
- Consolidate vitest config using test.projects pattern
- Add path aliases (@src, @tests) to simplify imports
- Fix tooltip positioning using actual handle element from shadow DOM
- Add context menu pre-selection based on current video playback rate
- Sync popup slider when video rate changes via native controls
- Add JSDoc documentation throughout
- Update test scripts: test runs all, test:unit runs unit only

+++ src/options.ts
@@ -12,22 +12,23 @@ import type Switch from '@ui5/webcomponents/dist/Switch.js';
 
 /**
  * Wrapper for the chrome.storage.sync for default settings.
- * @param {boolean} checkBoxState
- * @param {string} playbackRate
+ * @param {boolean} checkBoxState - Whether defaults are enabled
+ * @param {string} playbackRate - The playback rate value as string
  */
 async function saveDefaults(checkBoxState: boolean, playbackRate: string) {
 	await chrome.storage.sync.set({
 		defaults: {
 			enabled: checkBoxState,
-			playbackRate: Number.parseInt(playbackRate) // parse our string to a number
+			playbackRate: Number.parseFloat(playbackRate) // parseFloat to support fractional rates
 		}
 	});
 }
 
 /**
- *
- * @param defaultsCheckbox
- * @param defaultSpeedSelector
+ * Initialize the defaults from storage and populate UI elements.
+ * @param {CheckBox} defaultsCheckbox - The checkbox for enabling defaults
+ * @param {Select} defaultSpeedSelector - The dropdown for selecting default speed
+ * @returns {Promise<void>}
  */
 async function initDefaults(defaultsCheckbox: CheckBox, defaultSpeedSelector: Select): Promise<void> {
 	const { defaults } = <Defaults>await chrome.storage.sync.get('defaults');
